% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSIcalculator-package.R
\docType{package}
\name{OSIcalculator-package}
\alias{OSIcalculator}
\alias{OSIcalculator-package}
\title{OSIcalculator: Calculate the Obstruction Severity Index (OSI)}
\description{
Provides simple functions to calculate the Obstruction Severity Index (OSI) from lung function parameters such as FEV1 percent predicted and FEV1/FVC ratio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamLar27/OSIcalculator}
  \item Report bugs at \url{https://github.com/SamLar27/OSIcalculator/issues}
}

}
\author{
\strong{Maintainer}: Samuel Larouche \email{your@email.com}

}
\keyword{internal}
